requirements.txt
asgiref==3.4.1
autopep8==1.5.7
boto3==1.18.40
botocore==1.21.40
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
jmespath==0.10.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
